First update
Rough prototype, doesn't accept inputs yet but performs intended functions accurately. This prototype is capable of receiving one of a set of hard-coded recipes and outputting a list of required "basic" ingredients, compressed into blocks if possible.

Included files:
Main: A simple driver to run the system functions

MaterialCalculator: The primary functions which form the interface of the calculator. Currently, get_raw_mats receives a name and quantity of a recipe and returns a dictionary of raw material integers representing the minimum required inputs. Compress_inputs receives such a dictionary and compresses defined raw materials into blocks if they number more than 128.

Materials.py: Several basic dictionary shells and an array. basic_materials: the list of materials that the calculator will consider "raw," even if there are recipes for them in other dictionaries. RecipeDict: A (currently hardcoded) dictionary shell that contains recipes, with the key being the recipe output. CompressionRecipeDict: A RecipeDict containing recipes for compressing raw materials. The keys in this dictionary are the raw material, while the recipes are in format [compressed item, input quantity]

Recipe: An ADT containing minecraft recipes

WorkTree: Data types assisting the calculator in managing its workload and order. Dependencies and RecipeDependencies ensure that the quantity of any material is only calculated only after the quantity of all recipes that depend on it have already been calculated. DependencyDict is a shell for a dictionary holding dependencies whose key is the name of the item to be created. Together, these ensure that the tree of calculations resolves in the most efficient manner, and that recipes which produce multiple of their output don't overestimate the number of raw inputs required.